Rename Quilt Item Extension to Quilt Item Extensions No `provides` because this is 1.19.4
