--- library/item/item_extension/src/testmod/resources/quilt.mod.json
@@ -1,23 +1,0 @@
-{
-  "schema_version": 1,
-  "quilt_loader": {
-    "group": "org.quiltmc.qsl.item",
-    "id": "quilt_item_extension_testmod",
-    "version": "1.0.0",
-    "metadata": {
-      "name": "Quilt Item Extension API Test Mod",
-      "license": "Apache-2.0"
-    },
-    "intermediate_mappings": "net.fabricmc:intermediary",
-    "load_type": "always",
-    "entrypoints": {
-      "init": [
-        "org.quiltmc.qsl.item.extension.test.BowsTest"
-      ]
-    },
-    "depends": [
-      "quilt_loader",
-      "quilt_item_extension"
-    ]
-  }
-}
