+++ library/item/item_extensions/build.gradle
@@ -1,0 +1,22 @@
+plugins {
+	id("qsl.module")
+}
+
+qslModule {
+	name = "Quilt Item Extensions API"
+	moduleName = "item_extensions"
+	id = "quilt_item_extensions"
+	description = "An API for extending various vanilla items and providing relevant events for such items."
+	library = "item"
+	moduleDependencies {
+		core {
+			api("qsl_base")
+		}
+	}
+	injectedInterface("net/minecraft/class_1753") {
+		values = ["org/quiltmc/qsl/item/extensions/api/bow/BowExtensions"]
+	}
+	injectedInterface("net/minecraft/class_1764") {
+		values = ["org/quiltmc/qsl/item/extensions/api/crossbow/CrossbowExtensions"]
+	}
+}
