+++ library/item/item_extensions/src/main/java/org/quiltmc/qsl/item/extensions/mixin/bow/AbstractSkeletonEntityMixin.java
@@ -1,0 +1,44 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.item.extensions.mixin.bow;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+
+import net.minecraft.entity.EntityType;
+import net.minecraft.entity.ai.RangedAttackMob;
+import net.minecraft.entity.mob.AbstractSkeletonEntity;
+import net.minecraft.entity.mob.MobEntity;
+import net.minecraft.item.Item;
+import net.minecraft.item.ItemStack;
+import net.minecraft.world.World;
+
+import org.quiltmc.qsl.item.extensions.api.bow.BowExtensions;
+
+@Mixin(AbstractSkeletonEntity.class)
+public abstract class AbstractSkeletonEntityMixin extends MobEntity implements RangedAttackMob {
+	protected AbstractSkeletonEntityMixin(EntityType<? extends MobEntity> entityType, World world) {
+		super(entityType, world);
+	}
+
+	// Allows Skeletons to shoot custom bows by returning true
+	@Redirect(method = "updateAttackType", at = @At(value = "INVOKE", target = "Lnet/minecraft/item/ItemStack;isOf(Lnet/minecraft/item/Item;)Z"))
+	private boolean updateAttackType(ItemStack itemStack, Item item) {
+		return itemStack.getItem() instanceof BowExtensions;
+	}
+}
