+++ library/item/item_extensions/src/main/java/org/quiltmc/qsl/item/extensions/mixin/crossbow/CrossbowItemAccessor.java
@@ -1,0 +1,31 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.item.extensions.mixin.crossbow;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+import net.minecraft.item.CrossbowItem;
+import net.minecraft.item.ItemStack;
+
+@Mixin(CrossbowItem.class)
+public interface CrossbowItemAccessor {
+	@Invoker("getSpeed")
+	static float invokeGetSpeed(ItemStack stack) {
+		throw new IllegalStateException("Invoker injection failed.");
+	}
+}
