+++ library/item/item_extensions/src/main/java/org/quiltmc/qsl/item/extensions/mixin/crossbow/client/PlayerEntityRendererMixin.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2022 QuiltMC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.quiltmc.qsl.item.extensions.mixin.crossbow.client;
+
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
+
+import net.minecraft.client.network.AbstractClientPlayerEntity;
+import net.minecraft.client.render.entity.PlayerEntityRenderer;
+import net.minecraft.client.render.entity.model.BipedEntityModel;
+import net.minecraft.item.CrossbowItem;
+import net.minecraft.item.ItemStack;
+import net.minecraft.util.Hand;
+
+import org.quiltmc.qsl.item.extensions.api.crossbow.CrossbowExtensions;
+
+@Mixin(PlayerEntityRenderer.class)
+public class PlayerEntityRendererMixin {
+	@Inject(method = "getArmPose", at = @At("HEAD"), cancellable = true)
+	private static void getArmPose(AbstractClientPlayerEntity abstractClientPlayerEntity, Hand hand, CallbackInfoReturnable<BipedEntityModel.ArmPose> cir) {
+		ItemStack stackInHand = abstractClientPlayerEntity.getStackInHand(hand);
+
+		if (!abstractClientPlayerEntity.handSwinging && stackInHand.getItem() instanceof CrossbowExtensions && CrossbowItem.isCharged(stackInHand)) {
+			cir.setReturnValue(BipedEntityModel.ArmPose.CROSSBOW_HOLD);
+		}
+	}
+}
