+++ library/item/item_extensions/src/main/resources/assets/quilt_item_extensions/icon.png
