+++ library/item/item_extensions/src/main/resources/quilt_item_extensions.mixins.json
@@ -1,0 +1,24 @@
+{
+  "required": true,
+  "package": "org.quiltmc.qsl.item.extensions.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "ProjectileUtilMixin",
+    "bow.AbstractSkeletonEntityMixin",
+    "bow.BowAttackGoalMixin",
+    "bow.BowAttackMixin",
+    "bow.BowItemMixin",
+    "bow.EntityUseBowMixin",
+    "crossbow.CrossbowItemAccessor",
+    "crossbow.CrossbowItemMixin",
+    "crossbow.EntityUseCrossbowMixin"
+  ],
+  "client": [
+    "bow.client.SkeletonEntityModelMixin",
+    "client.HeldItemRendererMixin",
+    "crossbow.client.PlayerEntityRendererMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
