+++ library/item/item_extensions/src/testmod/resources/quilt.mod.json
@@ -1,0 +1,23 @@
+{
+  "schema_version": 1,
+  "quilt_loader": {
+    "group": "org.quiltmc.qsl.item",
+    "id": "quilt_item_extensions_testmod",
+    "version": "1.0.0",
+    "metadata": {
+      "name": "Quilt Item Extensions API Test Mod",
+      "license": "Apache-2.0"
+    },
+    "intermediate_mappings": "net.fabricmc:intermediary",
+    "load_type": "always",
+    "entrypoints": {
+      "init": [
+        "org.quiltmc.qsl.item.extensions.test.BowsTest"
+      ]
+    },
+    "depends": [
+      "quilt_loader",
+      "quilt_item_extensions"
+    ]
+  }
+}
